requirements-async.txt
# Async processing dependencies
aiofiles>=23.0.0
asyncio-throttle>=1.0.2

# Enhanced testing
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0

# Performance monitoring
memory-profiler>=0.60.0
psutil>=5.9.0